creation composant user-list-item
