+++ src/assets/scss/component/base-components/_base-button-option.scss
@@ -70,48 +70,6 @@ $base-button-option: 'base-button-option';
 
     &__remove {
       position: relative;
-
-      .delete__actions {
-        position: absolute;
-        right: 0;
-        bottom: 50%;
-        height: 41px;
-        width: 85px;
-        background: $white;
-        display: -ms-flexbox;
-        display: flex;
-        -ms-flex-align: center;
-        align-items: center;
-        -ms-flex-pack: distribute;
-        justify-content: space-around;
-        box-shadow: 0 2px 10px 0 rgba(0,0,0,.07);
-        transform: translateY(50%);
-        
-        .check,
-        .check-cross {
-          height: 28px;
-          width: 28px;
-          border-radius: 100%;
-          transition: all ease 0.25s;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-
-          &:hover {  
-            background: #F6F6F6;
-          }
-        }
-        .check {
-          .svg-icon {
-            fill: $success;
-          }
-        }
-        .check-cross {
-          .svg-icon {
-            fill: $danger;
-          }
-        }
-      }
     }
   }
 }
