+++ src/components/project/UsersList.vue
@@ -61,64 +61,7 @@
           </slot>
           <div class="users-list__body">
             <ul class="users-list__users">
-              <li v-for="(user, index) in filteredUsers" :key="index" class="users-list__user">
-                <div class="users-list__user__picture" v-if="user.hasAccepted">
-                  <img :src="user.photo" alt="" class="img-fluid circle" v-if="user.photo">
-                  <span class="user-menu__pic" v-else>{{ user.name|initialsFormat }}</span>
-                  <!--<span class="users-list__user__status users-list__user__status--inline" v-if="user.online"></span>
-                  <span class="users-list__user__status users-list__user__status--outline" v-else></span>-->
-                </div>
-                <div class="users-list__user__picture" v-else>
-                  <svgicon name="user-invit-pending" height="40" width="40"></svgicon>
-                </div>
-                <div class="users-list__user__datas" v-if="user.hasAccepted">
-                  <p>
-                    <span class="users-list__user__name">{{ user.name }}</span>
-                    <span>{{ user.job }}</span>
-                    <span>{{ user.company }}</span>
-                  </p>
-                  <span v-html="getBadge(user.role)"></span>
-                </div>
-                <div class="users-list__user__datas" v-else>
-                  <p>
-                    <span class="users-list__user__name">{{ user.name }}</span>
-                    <span>{{ $t('users.guest') }} - {{ $t('users.without_answer') }}. <a href="">{{ $t('users.resend_invitation') }}</a></span>
-                  </p>
-                </div>
-                <div class="users-list__user__actions" v-if="user.hasAccepted && isAdmin()">
-                  <base-button-option @option-toggled="toggleMenu" class="users-list__user__actions__menu" v-if="displayMenu">
-                    <ul>
-                        <li @click.stop.self="toggleRights()" :class="{'actif': displayRights}" class="arrow-left">
-                          {{ $t('users.modify_rights') }} <svgicon name="user-cadenas" width="12" height="12"></svgicon>
-
-                          <div class="new_folder_box" v-if="displayRights">
-                            <base-input-radio
-                              v-for="(right, indexRight) in rights"
-                              :key="indexRight"
-                              :id="user.id"
-                              :option="right"
-                              name="rights"
-                              @input="radioSelected"
-                              :selected="user.role"
-                            ></base-input-radio>
-                          </div>
-                      </li>
-                      <li @click.stop.self="showRemoveActions = true" class="base-button-option__menu__remove">
-                          {{ $t('users.remove') }} <svgicon name="user-croix" width="12" height="12"></svgicon>
-                          <transition name="slide-fade">
-                            <base-valid-delete v-if="showRemoveActions" @on-valid-action="removeUser(user.id)" @on-cancel-action="showRemoveActions = false"></base-valid-delete>
-                          </transition>
-                        </li>
-                    </ul>
-                  </base-button-option>
-                  <div v-else class="base-button-option__menu__remove">
-                    <svgicon name="delete" height="15" width="18" @click="showRemoveActions = true"></svgicon>
-                    <transition name="slide-fade">
-                      <base-valid-delete v-if="showRemoveActions" @on-valid-action="removeUser(user.id)" @on-cancel-action="showRemoveActions = false"></base-valid-delete>
-                    </transition>
-                  </div>
-                </div>
-              </li>
+              <users-list-item v-for="(user, index) in filteredUsers" :key="index" :user="user" :displayMenu="displayMenu"></users-list-item>
             </ul>
           </div>
         </div>
@@ -130,8 +73,8 @@ import BaseCard from '@/components/base-components/BaseCard'
 import BaseButtonOption from '@/components/base-components/BaseButtonOption'
 import BaseValidDelete from '@/components/base-components/BaseValidDelete'
 import BaseInputRadio from '@/components/base-components/BaseInputRadio'
+import UsersListItem from '@/components/project/UsersListItem'
 import { mixin as clickaway } from 'vue-clickaway'
-import _ from 'lodash'
 
 export default {
   mixins: [ clickaway ],
@@ -139,7 +82,8 @@ export default {
     BaseCard,
     BaseButtonOption,
     BaseInputRadio,
-    BaseValidDelete
+    BaseValidDelete,
+    UsersListItem
   },
   props: {
     displayMenu: {
@@ -156,9 +100,7 @@ export default {
   },
   data () {
     return {
-      displayRights: false,
       displayRightsInvitation: false,
-      showRemoveActions: false,
       displaySendInvit: false,
       displaySearchUser: false,
       clicked: false,
@@ -169,20 +111,6 @@ export default {
       },
       mailInvitation: '',
       rightInvitation: null,
-      rights: [
-        {
-          text: this.$t('users.administrator'),
-          value: 100
-        },
-        {
-          text: this.$t('users.user'),
-          value: 50
-        },
-        {
-          text: this.$t('users.guest'),
-          value: 25
-        }
-      ],
       guests: [
         {
           id: 7,
@@ -198,16 +126,6 @@ export default {
     }
   },
   methods: {
-    toggleRights () {
-      this.displayRights = !this.displayRights
-      this.showRemoveActions = false
-    },
-    toggleMenu (isOpened) {
-      if (!isOpened) {
-        this.showRemoveActions = false
-        this.displayRights = false
-      }
-    },
     radioSelected (object) {
       console.log('object', object)
       // Call ajax
@@ -215,19 +133,6 @@ export default {
     removeUser (userId) {
       console.log(userId)
       // Call to remove user
-    },
-    isAdmin () {
-      if (this.$store.state.project.selectedProject !== null) {
-        if (this.$store.state.project.selectedProject.role === 100) {
-          return true
-        }
-      } else {
-        if (this.$store.state.currentCloud.role === 100) {
-          return true
-        }
-      }
-
-      return false
     },
     openSendInvit () {
       this.clicked = false
@@ -258,18 +163,6 @@ export default {
     resetSearchUser () {
       this.displaySearchUser = false
       this.searchFilter = ''
-    },
-    getBadge (role) {
-      let roleText = _.find(this.rights, function (r) {
-        return r.value === role
-      }).text
-      if (role === 100) {
-        return '<span class="badge badge-primary">' + roleText + '</span>'
-      } else if (role === 50) {
-        return '<span class="badge badge-success">' + roleText + '</span>'
-      } else {
-        return '<span class="badge badge-secondary text-white">' + roleText + '</span>'
-      }
     }
   },
   created () {
